requirements.txt
# Requirements for Vietnamese Speech-to-Text Dataset EDA
# Các thư viện cần thiết cho phân tích EDA dataset speech-to-text tiếng Việt

# Core data analysis libraries
pandas>=1.5.0
numpy>=1.21.0

# Visualization libraries
matplotlib>=3.5.0
seaborn>=0.11.0

# Text processing and NLP
scikit-learn>=1.1.0
wordcloud>=1.9.0

# Audio processing for Whisper fine-tuning
librosa>=0.9.0
soundfile>=0.12.0

# Whisper and OpenAI libraries
openai-whisper>=20231117
tiktoken>=0.5.0

# Additional audio processing
pydub>=0.25.1

# Additional utilities
pathlib2>=2.3.0  # For older Python versions compatibility

# Optional: For better Vietnamese text processing
# underthesea>=1.3.0  # Vietnamese NLP toolkit (uncomment if needed)

# YouTube crawling and processing
yt-dlp>=2023.7.6
requests>=2.31.0
beautifulsoup4>=4.12.0
pysrt>=1.1.2
pydub>=0.25.1

# Web automation for subtitle downloading
selenium>=4.15.2
webdriver-manager>=4.0.1

# Development and testing (optional)
# jupyter>=1.0.0
# ipykernel>=6.0.0

# Knowledge Distillation and Dataset Processing
# PyTorch 2.7.1 with CUDA 12.8 for RTX 50 series (Blackwell architecture) support
# Install with: python -m pip install -U torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
torch>=2.7.1
torchaudio>=2.7.1
torchvision>=0.22.1
torchcodec>=0.1.0
transformers>=4.40.0  # Required for PyTorch 2.9.0 and PhoWhisper support
datasets>=2.14.0
python-dotenv>=1.0.0
tqdm>=4.65.0
accelerate>=0.27.0  # Required for mixed precision training with PyTorch 2.9.0
evaluate>=0.4.0

# TensorFlow with GPU support for RTX 5000 series
# Use nightly build with auto CUDA support (works with RTX 5060)
# Install with: pip install tf-nightly[and-cuda]
tf-nightly[and-cuda]  # TensorFlow nightly with CUDA bundled

# YAML support for configs
pyyaml>=6.0